src/main/java/com/vahner/airticketsapp/service/impl/UserDetailServiceImpl.java
package com.vahner.airticketsapp.service.impl;

import com.vahner.airticketsapp.entity.AppUser;
import com.vahner.airticketsapp.repository.AppUserRepository;
import com.vahner.airticketsapp.security.Authentication;
import lombok.AllArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@AllArgsConstructor
public class UserDetailServiceImpl implements UserDetailsService {

    private final AppUserRepository appUserRepository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        AppUser appUser = appUserRepository.findByUsername(username)
                .orElseThrow(() -> new UsernameNotFoundException("User not found with username: " + username));
        return new Authentication(appUser);
    }
}